+++ app/models/group_lesson_summary.rb
@@ -6,14 +6,13 @@
 #  grade_count        :bigint
 #  group_chapter_name :text
 #  lesson_date        :date
-#  lesson_uid         :uuid             primary key
 #  chapter_id         :integer
 #  group_id           :integer
-#  lesson_id          :integer
+#  lesson_id          :uuid             primary key
 #  subject_id         :integer
 #
 class GroupLessonSummary < ApplicationRecord
-  self.primary_key = :lesson_uid
+  self.primary_key = :lesson_id
   belongs_to :chapter
 
   def readonly?
@@ -29,11 +28,11 @@ class GroupLessonSummary < ApplicationRecord
   def self.select_for_graphing(relation, sorting_direction, number_of_elements)
     relation
       .order(lesson_date: sorting_direction).limit(number_of_elements)
-      .select(Arel.sql('extract(EPOCH FROM lesson_date) :: INT AS timestamp'), :average_mark, :lesson_date, :lesson_uid, :lesson_id)
+      .select(Arel.sql('extract(EPOCH FROM lesson_date) :: INT AS timestamp'), :average_mark, :lesson_date, :lesson_id)
   end
 
   def self.process_around_result(lesson, result, number_of_elements)
-    index = result.index { |l| l.lesson_uid == lesson.uid }
+    index = result.index { |l| l.lesson_id == lesson.id }
     index = find_closest_index(result, lesson) if index.nil?
     number_of_elements = [result.size, number_of_elements].min
     return [] if result == []