+++ spec/factories/absences.rb
@@ -2,9 +2,10 @@
 #
 # Table name: absences
 #
-#  id         :integer          not null, primary key
-#  lesson_id  :integer          not null
-#  student_id :integer          not null
+#  id            :integer          not null, primary key
+#  lesson_id     :uuid             not null
+#  lesson_old_id :integer          not null
+#  student_id    :integer          not null
 #
 # Indexes
 #
@@ -13,12 +14,11 @@
 #
 # Foreign Keys
 #
-#  fk_rails_...  (lesson_id => lessons.id)
-#  fk_rails_...  (student_id => students.id)
+#  absences_lesson_id_fk  (lesson_id => lessons.id)
+#  fk_rails_...           (student_id => students.id)
 #
 FactoryBot.define do
   factory :absence do
-    student { create :student }
-    lesson { create :lesson }
+    lesson_old_id { lesson.old_id }
   end
 end