+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Restructured Lessons to go by their UUID
 - Disabled decimals for 'No of Assessments' & 'No. of Lessons' in analytics
 - Upgraded Rails to 7.1
 - Enabled YJIT